#23 Moved the tests in the test folder
